Move bulk of logic to controllers
