+++ vaulty-mail/server/src/routes.rs
@@ -1,6 +1,6 @@
-use vaulty::{email, mailgun};
-
 use warp::{Filter, Rejection, reply::Reply};
+
+use super::controllers;
 
 pub fn index() -> impl Filter<Extract = (&'static str, ), Error = Rejection> + Clone {
     // GET /hello/warp => 200 OK with body "Hello, warp!"
@@ -20,36 +20,5 @@ pub fn mailgun() -> impl Filter<Extract = (impl Reply, ), Error = Rejection> + C
                     .map_err(|_e| warp::reject::not_found())
             }
         }))
-        .and_then(|content_type: Option<String>, body: String| {
-            async move {
-                if let None = content_type {
-                    return Err(warp::reject::not_found());
-                }
-
-                let mut mail = match mailgun::Email::from_body(&body, &content_type.unwrap()) {
-                    Ok(m) => m,
-                    Err(e) => {
-                        log::error!("{:?}", e);
-                        return Err(warp::reject::not_found());
-                    }
-                };
-
-                if let Err(_e) = mail.fetch_attachments().await {
-                    return Err(warp::reject::not_found());
-                };
-
-                log::info!("Fetched all attachments successfully!");
-
-                let handler = vaulty::EmailHandler::new();
-                let mail: email::Email = mail.into();
-
-                if let Err(_e) = handler.handle(mail).await {
-                    return Err(warp::reject::not_found());
-                }
-
-                log::info!("Mail handling completed");
-
-                Ok(warp::reply())
-            }
-        })
+        .and_then(controllers::mailgun)
 }
